requirements.txt
torchvision~=0.12.0
numpy~=1.22.3
matplotlib~=3.5.1
Pillow~=9.0.1

tqdm~=4.64.0